Refactor edge server into lib; fix build warnings for 3.2.4.
Move HTTP handler tree from main.rs to server.rs so the prototype bin is
thin and clippy/tests see one module graph. Fix duplicate json-api route
and clippy redundant-closure/useless-format issues without allow attrs.
